backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
supabase==2.3.0

# Caching and message queue
redis==5.0.1
celery==5.3.4

# Authentication and security
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# HTTP client and API integration
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# AI and ML libraries
langchain==0.0.335
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.3.7
anthropic==0.7.8
crewai==0.1.0
crewai-tools==0.1.0

# Data processing
pandas==2.1.4
numpy==1.25.2
polars==0.20.2

# Validation and schemas
jsonschema==4.20.0
zod==0.0.1

# Monitoring and observability
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.38.0

# Scheduling and cron
apscheduler==3.10.4
croniter==2.0.1

# File processing and utilities
python-magic==0.4.27
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3

# Database drivers for new agents
psycopg2-binary==2.9.9
pymysql==1.1.0

# File handling and email
aiofiles==23.2.1
dnspython==2.4.2

# Additional data processing
openpyxl==3.1.2
xlsxwriter==3.1.9